requirements.txt
pyqt5
dnspython
requests
python-nmap
tkinter
cryptography
psutil
